Split typechecker behavior validation
