+++ src/typechecker/mod.rs
@@ -8,6 +8,9 @@
 //! The typechecker NEVER defaults unknown types to I32. If a type can't be
 //! resolved, it's an error.
 
+mod behavior_associations;
+mod behavior_impl_support;
+mod behavior_impl_validation;
 mod closures;
 mod environment;
 mod expressions;
@@ -2860,1030 +2863,6 @@ impl TypeChecker {
         }
 
         Some(ast_substitutions)
-    }
-
-    fn check_behavior_requires(
-        &mut self,
-        type_name: &str,
-        behavior: &str,
-        behavior_type_args: &[AstType],
-        span: Span,
-    ) {
-        let resolver_required_ref = self.resolver_required_ref_for(type_name, behavior);
-        if self.should_skip_missing_resolver_behavior_ref(
-            resolver_required_ref.as_ref(),
-            type_name,
-            &self.resolver_missing_behavior_required_refs,
-        ) {
-            return;
-        }
-        let (behavior, behavior_type_args) =
-            Self::behavior_ref_parts(resolver_required_ref.as_ref(), behavior, behavior_type_args);
-
-        if !self.structs.contains_key(type_name) && !self.enums.contains_key(type_name) {
-            self.diagnostics.push(Diagnostic::error(
-                "E6005",
-                format!("undefined type `{}`", type_name),
-                span,
-            ));
-            return;
-        }
-
-        if self.reject_unspecialized_generic_type(type_name, span) {
-            return;
-        }
-
-        let Some(_) = self.behavior_type_arg_substitutions(
-            behavior,
-            behavior_type_args,
-            &HashSet::new(),
-            span,
-        ) else {
-            return;
-        };
-        let behavior_key = self.behavior_reference_key(behavior, behavior_type_args);
-
-        if !self.type_implements_behavior(type_name, &behavior_key) {
-            self.diagnostics.push(Diagnostic::error(
-                "E6007",
-                format!(
-                    "type `{}` does not implement required behavior `{}`",
-                    type_name, behavior_key
-                ),
-                span,
-            ));
-        }
-    }
-
-    fn resolver_required_ref_for(
-        &mut self,
-        type_name: &str,
-        behavior: &str,
-    ) -> Option<BehaviorRefMetadata> {
-        self.resolver_behavior_ref_for(BehaviorRefRole::Required, type_name, behavior)
-    }
-
-    fn check_behavior_extends(
-        &mut self,
-        behavior: &str,
-        parent: &str,
-        parent_type_args: &[AstType],
-        span: Span,
-    ) {
-        if !self.behaviors.contains_key(behavior) {
-            self.diagnostics.push(Diagnostic::error(
-                "E6006",
-                format!("undefined behavior `{}`", behavior),
-                span,
-            ));
-            return;
-        }
-
-        let scoped_type_params: HashSet<String> = self
-            .behaviors
-            .get(behavior)
-            .map(|info| info.type_params.iter().cloned().collect())
-            .unwrap_or_default();
-        let Some(_) = self.behavior_type_arg_substitutions(
-            parent,
-            parent_type_args,
-            &scoped_type_params,
-            span,
-        ) else {
-            return;
-        };
-
-        let parent_ref = self.behavior_parent_ref(parent, parent_type_args);
-        let parent_display = behavior_ref_display(parent, parent_type_args);
-        let parents = self
-            .behavior_extends
-            .entry(behavior.to_string())
-            .or_default();
-        if parents
-            .iter()
-            .any(|existing| existing.key == parent_ref.key)
-        {
-            self.diagnostics.push(Diagnostic::error(
-                "E6011",
-                format!("duplicate behavior inheritance `{behavior}.extends({parent_display})`"),
-                span,
-            ));
-            return;
-        }
-
-        parents.push(parent_ref);
-        self.behavior_extends_spans
-            .entry(behavior.to_string())
-            .or_insert(span);
-    }
-
-    fn validate_behavior_extends_cycles(&mut self) {
-        let behaviors: Vec<String> = self.behavior_extends.keys().cloned().collect();
-        for behavior in behaviors {
-            let mut visiting = HashSet::new();
-            let mut visited = HashSet::new();
-            if self.behavior_extends_has_cycle(&behavior, &mut visiting, &mut visited) {
-                let span = self
-                    .behavior_extends_spans
-                    .get(&behavior)
-                    .copied()
-                    .unwrap_or_else(Span::dummy);
-                self.diagnostics.push(Diagnostic::error(
-                    "E6008",
-                    format!("behavior inheritance cycle involving `{}`", behavior),
-                    span,
-                ));
-            }
-        }
-    }
-
-    fn behavior_extends_has_cycle(
-        &self,
-        behavior: &str,
-        visiting: &mut HashSet<String>,
-        visited: &mut HashSet<String>,
-    ) -> bool {
-        if visiting.contains(behavior) {
-            return true;
-        }
-        if !visited.insert(behavior.to_string()) {
-            return false;
-        }
-
-        visiting.insert(behavior.to_string());
-        let has_cycle = self.behavior_extends.get(behavior).is_some_and(|parents| {
-            parents
-                .iter()
-                .any(|parent| self.behavior_extends_has_cycle(&parent.key, visiting, visited))
-        });
-        visiting.remove(behavior);
-        has_cycle
-    }
-
-    fn validate_behavior_method_coherence(&mut self) {
-        let behaviors: Vec<String> = self.behavior_extends.keys().cloned().collect();
-        let mut diagnostics = Vec::new();
-
-        for behavior in behaviors {
-            let mut seen_behaviors = HashSet::new();
-            let mut seen_methods = HashMap::new();
-            self.collect_behavior_method_coherence_errors(
-                &behavior,
-                &behavior,
-                &HashMap::new(),
-                &mut seen_behaviors,
-                &mut seen_methods,
-                &mut diagnostics,
-            );
-        }
-
-        self.diagnostics.extend(diagnostics);
-    }
-
-    fn collect_behavior_method_coherence_errors(
-        &self,
-        behavior: &str,
-        root_behavior: &str,
-        substitutions: &HashMap<String, AstType>,
-        seen_behaviors: &mut HashSet<String>,
-        seen_methods: &mut HashMap<String, ast::BehaviorMethod>,
-        diagnostics: &mut Vec<Diagnostic>,
-    ) {
-        if !self.mark_behavior_seen(behavior, substitutions, seen_behaviors) {
-            return;
-        }
-
-        if let Some(parents) = self.behavior_extends.get(behavior) {
-            for parent in parents {
-                let parent_substitutions =
-                    self.behavior_parent_type_param_substitutions(parent, substitutions);
-                self.collect_behavior_method_coherence_errors(
-                    &parent.behavior,
-                    root_behavior,
-                    &parent_substitutions,
-                    seen_behaviors,
-                    seen_methods,
-                    diagnostics,
-                );
-            }
-        }
-
-        if let Some(info) = self.behaviors.get(behavior) {
-            for method in &info.methods {
-                let method = substituted_behavior_method_signature(method, substitutions);
-
-                if let Some(previous) = seen_methods.get(&method.name) {
-                    if !behavior_method_signatures_match(previous, &method) {
-                        diagnostics.push(Diagnostic::error(
-                            "E6009",
-                            format!(
-                                "conflicting behavior method `{}` inherited by `{}`",
-                                method.name, root_behavior
-                            ),
-                            method.span,
-                        ));
-                    }
-                } else {
-                    seen_methods.insert(method.name.clone(), method);
-                }
-            }
-        }
-    }
-
-    fn type_implements_behavior(&self, type_name: &str, behavior: &str) -> bool {
-        if self
-            .behavior_impls
-            .contains(&(type_name.to_string(), behavior.to_string()))
-        {
-            return true;
-        }
-
-        self.behavior_impls
-            .iter()
-            .any(|(implemented_type, implemented_behavior)| {
-                implemented_type == type_name
-                    && self.behavior_inherits_from(implemented_behavior, behavior)
-            })
-    }
-
-    fn behavior_inherits_from(&self, behavior: &str, parent: &str) -> bool {
-        self.behavior_inherits_from_inner(behavior, parent, &mut HashSet::new())
-    }
-
-    fn behavior_inherits_from_inner(
-        &self,
-        behavior: &str,
-        parent: &str,
-        seen: &mut HashSet<String>,
-    ) -> bool {
-        if !seen.insert(behavior.to_string()) {
-            return false;
-        }
-
-        if self.behavior_extends_parent_matches(behavior, &HashMap::new(), parent, seen) {
-            return true;
-        }
-
-        self.behavior_refs_by_key
-            .get(behavior)
-            .is_some_and(|behavior_ref| {
-                let substitutions = self.behavior_type_param_substitutions(
-                    &behavior_ref.behavior,
-                    &behavior_ref.type_args,
-                );
-                self.behavior_extends_parent_matches(
-                    &behavior_ref.behavior,
-                    &substitutions,
-                    parent,
-                    seen,
-                )
-            })
-    }
-
-    fn behavior_extends_parent_matches(
-        &self,
-        behavior: &str,
-        substitutions: &HashMap<String, AstType>,
-        parent: &str,
-        seen: &mut HashSet<String>,
-    ) -> bool {
-        self.behavior_extends.get(behavior).is_some_and(|parents| {
-            parents.iter().any(|candidate| {
-                let candidate_args: Vec<AstType> = candidate
-                    .type_args
-                    .iter()
-                    .map(|type_arg| substitute_behavior_ast_type(type_arg, substitutions))
-                    .collect();
-                let candidate_ref = self.behavior_parent_ref(&candidate.behavior, &candidate_args);
-                candidate_ref.key == parent
-                    || self.behavior_ref_inherits_from_inner(&candidate_ref, parent, seen)
-            })
-        })
-    }
-
-    fn behavior_ref_inherits_from_inner(
-        &self,
-        behavior_ref: &BehaviorParentRef,
-        parent: &str,
-        seen: &mut HashSet<String>,
-    ) -> bool {
-        if !seen.insert(behavior_ref.key.clone()) {
-            return false;
-        }
-
-        let substitutions =
-            self.behavior_type_param_substitutions(&behavior_ref.behavior, &behavior_ref.type_args);
-        self.behavior_extends_parent_matches(&behavior_ref.behavior, &substitutions, parent, seen)
-    }
-
-    fn check_behavior_impl(
-        &mut self,
-        type_name: &str,
-        behavior: &str,
-        behavior_type_args: &[AstType],
-        methods: &[Declaration],
-        span: Span,
-        symbols: Option<&SymbolTable>,
-    ) {
-        let resolver_impl_ref = self.resolver_impl_ref_for(type_name, behavior);
-        if self.should_skip_missing_resolver_behavior_ref(
-            resolver_impl_ref.as_ref(),
-            type_name,
-            &self.resolver_missing_behavior_impl_refs,
-        ) {
-            return;
-        }
-        let (behavior, behavior_type_args) =
-            Self::behavior_ref_parts(resolver_impl_ref.as_ref(), behavior, behavior_type_args);
-
-        if !self.structs.contains_key(type_name) && !self.enums.contains_key(type_name) {
-            self.diagnostics.push(Diagnostic::error(
-                "E6005",
-                format!("undefined type `{}`", type_name),
-                span,
-            ));
-            return;
-        }
-
-        if self.reject_unspecialized_generic_type(type_name, span) {
-            return;
-        }
-
-        let Some(behavior_substitutions) = self.behavior_type_arg_substitutions(
-            behavior,
-            behavior_type_args,
-            &HashSet::new(),
-            span,
-        ) else {
-            return;
-        };
-        let behavior_key = self.behavior_reference_key(behavior, behavior_type_args);
-
-        if self
-            .behavior_impls
-            .contains(&(type_name.to_string(), behavior_key.clone()))
-        {
-            self.diagnostics.push(Diagnostic::error(
-                "E6003",
-                format!(
-                    "duplicate implementation of behavior `{}` for type `{}`",
-                    behavior_key, type_name
-                ),
-                span,
-            ));
-            return;
-        }
-
-        if let Some(existing) = self.find_overlapping_behavior_impl(type_name, &behavior_key) {
-            self.diagnostics.push(Diagnostic::error(
-                "E6010",
-                format!(
-                    "overlapping implementations of behaviors `{}` and `{}` for type `{}`",
-                    existing, behavior_key, type_name
-                ),
-                span,
-            ));
-            return;
-        }
-
-        self.behavior_impls
-            .insert((type_name.to_string(), behavior_key.clone()));
-        self.behavior_refs_by_key.insert(
-            behavior_key.clone(),
-            self.behavior_parent_ref(behavior, behavior_type_args),
-        );
-        let required_methods =
-            self.behavior_methods_for_impl(behavior, &behavior_substitutions, &mut HashSet::new());
-        let mut unmatched_required: VecDeque<String> = required_methods
-            .iter()
-            .map(|required| required.name.clone())
-            .collect();
-        let effective_methods = self.effective_behavior_impl_methods(
-            symbols,
-            type_name,
-            methods,
-            &mut unmatched_required,
-        );
-
-        for method in &effective_methods {
-            if let Declaration::Function { span, .. } = method.declaration {
-                if !required_methods
-                    .iter()
-                    .any(|required| required.name == method.method_name)
-                {
-                    self.diagnostics.push(Diagnostic::error(
-                        "E6005",
-                        format!(
-                            "method `{}` is not declared by behavior `{}`",
-                            method.method_name, behavior_key
-                        ),
-                        *span,
-                    ));
-                }
-            }
-        }
-
-        for required in &required_methods {
-            let Some(actual) =
-                effective_methods
-                    .iter()
-                    .find_map(|method| match method.declaration {
-                        Declaration::Function {
-                            params,
-                            return_type,
-                            span,
-                            ..
-                        } if method.method_name == required.name => {
-                            Some((params, return_type, *span))
-                        }
-                        _ => None,
-                    })
-            else {
-                if required.default_body.is_some() {
-                    continue;
-                }
-                self.diagnostics.push(Diagnostic::error(
-                    "E6001",
-                    format!(
-                        "type `{}` implementation of `{}` is missing required method `{}`",
-                        type_name, behavior_key, required.name
-                    ),
-                    span,
-                ));
-                continue;
-            };
-
-            let (actual_params, actual_return_type, actual_span) = actual;
-            let collected_signature =
-                self.resolver_backed_method_signature(type_name, &required.name);
-            let actual_param_types: Vec<AstType> = collected_signature
-                .map(|info| {
-                    info.params
-                        .iter()
-                        .map(|(_, ty)| ty.clone())
-                        .collect::<Vec<_>>()
-                })
-                .unwrap_or_else(|| actual_params.iter().map(|param| param.ty.clone()).collect());
-            let actual_return = collected_signature
-                .map(|info| info.return_type.clone())
-                .unwrap_or_else(|| actual_return_type.clone().unwrap_or(AstType::Void));
-
-            if actual_param_types.len() != required.params.len() {
-                self.diagnostics.push(Diagnostic::error(
-                    "E6002",
-                    format!(
-                        "method `{}` for behavior `{}` expects {} parameters, found {}",
-                        required.name,
-                        behavior_key,
-                        required.params.len(),
-                        actual_param_types.len()
-                    ),
-                    actual_span,
-                ));
-                continue;
-            }
-
-            for (idx, (expected, actual_ty)) in required
-                .params
-                .iter()
-                .zip(actual_param_types.iter())
-                .enumerate()
-            {
-                if !self.impl_ast_types_compatible(&expected.ty, actual_ty, type_name) {
-                    self.diagnostics.push(Diagnostic::error(
-                        "E6002",
-                        format!(
-                            "parameter {} for method `{}` in behavior `{}` expects `{}`, found `{}`",
-                            idx + 1,
-                            required.name,
-                            behavior_key,
-                            self.impl_type_display(&expected.ty, type_name),
-                            actual_ty.display_name()
-                        ),
-                        actual_params
-                            .get(idx)
-                            .map(|param| param.span)
-                            .unwrap_or(actual_span),
-                    ));
-                }
-            }
-
-            let expected_return = required.return_type.as_ref().unwrap_or(&AstType::Void);
-            if !self.impl_ast_types_compatible(expected_return, &actual_return, type_name) {
-                self.diagnostics.push(Diagnostic::error(
-                    "E6002",
-                    format!(
-                        "method `{}` for behavior `{}` expects return `{}`, found `{}`",
-                        required.name,
-                        behavior_key,
-                        self.impl_type_display(expected_return, type_name),
-                        actual_return.display_name()
-                    ),
-                    actual_span,
-                ));
-            }
-        }
-    }
-
-    fn resolver_impl_ref_for(
-        &mut self,
-        type_name: &str,
-        behavior: &str,
-    ) -> Option<BehaviorRefMetadata> {
-        self.resolver_behavior_ref_for(BehaviorRefRole::Impl, type_name, behavior)
-    }
-
-    fn resolver_behavior_ref_for(
-        &mut self,
-        role: BehaviorRefRole,
-        type_name: &str,
-        behavior: &str,
-    ) -> Option<BehaviorRefMetadata> {
-        match role {
-            BehaviorRefRole::Impl => Self::pop_resolver_behavior_ref(
-                self.resolver_backed_collection,
-                &mut self.resolver_behavior_impl_refs,
-                type_name,
-                behavior,
-            ),
-            BehaviorRefRole::Required => Self::pop_resolver_behavior_ref(
-                self.resolver_backed_collection,
-                &mut self.resolver_behavior_required_refs,
-                type_name,
-                behavior,
-            ),
-            BehaviorRefRole::Parent => None,
-        }
-    }
-
-    fn behavior_ref_parts<'a>(
-        resolver_ref: Option<&'a BehaviorRefMetadata>,
-        behavior: &'a str,
-        behavior_type_args: &'a [AstType],
-    ) -> (&'a str, &'a [AstType]) {
-        resolver_ref
-            .map(|reference| (reference.name.as_str(), reference.type_args.as_slice()))
-            .unwrap_or((behavior, behavior_type_args))
-    }
-
-    fn pop_resolver_behavior_ref(
-        resolver_backed_collection: bool,
-        refs_by_type: &mut HashMap<String, VecDeque<BehaviorRefMetadata>>,
-        type_name: &str,
-        behavior: &str,
-    ) -> Option<BehaviorRefMetadata> {
-        if !resolver_backed_collection {
-            return None;
-        }
-
-        let refs = refs_by_type.get_mut(type_name)?;
-        Self::pop_resolver_behavior_ref_from_queue(refs, behavior)
-    }
-
-    fn should_skip_missing_resolver_behavior_ref(
-        &self,
-        resolver_ref: Option<&BehaviorRefMetadata>,
-        type_name: &str,
-        missing_refs: &HashSet<String>,
-    ) -> bool {
-        self.resolver_backed_collection
-            && resolver_ref.is_none()
-            && missing_refs.contains(type_name)
-    }
-
-    fn pop_resolver_behavior_ref_from_queue(
-        refs: &mut VecDeque<BehaviorRefMetadata>,
-        behavior: &str,
-    ) -> Option<BehaviorRefMetadata> {
-        let index = Self::resolver_behavior_ref_queue_index(refs, behavior)?;
-        refs.remove(index)
-    }
-
-    fn resolver_behavior_impl_ref_for_peek(
-        &self,
-        type_name: &str,
-        behavior: &str,
-    ) -> Option<&BehaviorRefMetadata> {
-        Self::peek_resolver_behavior_ref(
-            self.resolver_backed_collection,
-            &self.resolver_behavior_impl_refs,
-            type_name,
-            behavior,
-        )
-    }
-
-    fn peek_resolver_behavior_ref<'a>(
-        resolver_backed_collection: bool,
-        refs_by_type: &'a HashMap<String, VecDeque<BehaviorRefMetadata>>,
-        type_name: &str,
-        behavior: &str,
-    ) -> Option<&'a BehaviorRefMetadata> {
-        if !resolver_backed_collection {
-            return None;
-        }
-
-        let refs = refs_by_type.get(type_name)?;
-        Self::resolver_behavior_ref_queue_index(refs, behavior).and_then(|index| refs.get(index))
-    }
-
-    fn resolver_behavior_ref_queue_index(
-        refs: &VecDeque<BehaviorRefMetadata>,
-        behavior: &str,
-    ) -> Option<usize> {
-        Self::named_queue_index(refs, behavior, |reference| reference.name.as_str())
-    }
-
-    fn named_queue_index<T>(
-        items: &VecDeque<T>,
-        name: &str,
-        item_name: impl Fn(&T) -> &str,
-    ) -> Option<usize> {
-        items
-            .iter()
-            .position(|item| item_name(item) == name)
-            .or_else(|| (!items.is_empty()).then_some(0))
-    }
-
-    fn named_queue_index_preserving_future_front<'a, T>(
-        items: &VecDeque<T>,
-        name: &str,
-        future_names: impl IntoIterator<Item = &'a str>,
-        item_name: impl Fn(&T) -> &str,
-    ) -> Option<usize> {
-        if let Some(index) = items.iter().position(|item| item_name(item) == name) {
-            return Some(index);
-        }
-
-        let front_name = item_name(items.front()?);
-        (!future_names
-            .into_iter()
-            .any(|future_name| future_name == front_name))
-        .then_some(0)
-    }
-
-    fn resolver_behavior_impl_ref_parts<'a>(
-        &'a self,
-        type_name: &str,
-        behavior: &'a str,
-        behavior_type_args: &'a [AstType],
-    ) -> (&'a str, &'a [AstType]) {
-        match self.resolver_behavior_impl_ref_for_peek(type_name, behavior) {
-            Some(implementation) => (
-                implementation.name.as_str(),
-                implementation.type_args.as_slice(),
-            ),
-            None => (behavior, behavior_type_args),
-        }
-    }
-
-    fn find_overlapping_behavior_impl(&self, type_name: &str, behavior: &str) -> Option<String> {
-        self.behavior_impls
-            .iter()
-            .filter(|(implemented_type, _)| implemented_type == type_name)
-            .map(|(_, implemented_behavior)| implemented_behavior)
-            .find(|implemented_behavior| {
-                self.behavior_inherits_from(implemented_behavior, behavior)
-                    || self.behavior_inherits_from(behavior, implemented_behavior)
-            })
-            .cloned()
-    }
-
-    fn reject_unspecialized_generic_type(&mut self, type_name: &str, span: Span) -> bool {
-        let type_param_count = self
-            .structs
-            .get(type_name)
-            .map(|info| info.type_params.len())
-            .or_else(|| self.enums.get(type_name).map(|info| info.type_params.len()))
-            .unwrap_or(0);
-        if type_param_count == 0 {
-            return false;
-        }
-
-        self.diagnostics.push(Diagnostic::error(
-            "E6013",
-            format!(
-                "generic type `{}` expects {} type arguments, found 0",
-                type_name, type_param_count
-            ),
-            span,
-        ));
-        true
-    }
-
-    fn behavior_default_methods_for_impl(
-        &self,
-        type_name: &str,
-        behavior: &str,
-        behavior_type_args: &[AstType],
-        methods: &[Declaration],
-    ) -> Vec<DefaultBehaviorMethod> {
-        let behavior_substitutions =
-            self.behavior_type_param_substitutions(behavior, behavior_type_args);
-        self.behavior_methods_for_impl(behavior, &behavior_substitutions, &mut HashSet::new())
-            .iter()
-            .filter(|required| {
-                required.default_body.is_some()
-                    && !self.impl_methods_include_behavior_method(
-                        type_name,
-                        methods,
-                        &required.name,
-                    )
-            })
-            .filter_map(|required| {
-                let body = required.default_body.clone()?;
-                Some(DefaultBehaviorMethod {
-                    name: required.name.clone(),
-                    params: required
-                        .params
-                        .iter()
-                        .map(|param| Param {
-                            name: param.name.clone(),
-                            ty: concrete_self_ast_type(&param.ty, type_name),
-                            mutable: param.mutable,
-                            span: param.span,
-                        })
-                        .collect(),
-                    return_type: required
-                        .return_type
-                        .as_ref()
-                        .map(|ty| concrete_self_ast_type(ty, type_name)),
-                    body,
-                    span: required.span,
-                })
-            })
-            .collect()
-    }
-
-    fn seed_behavior_default_method_signature(
-        &mut self,
-        type_name: &str,
-        default: &DefaultBehaviorMethod,
-    ) {
-        let key = Self::method_key(type_name, &default.name);
-        self.methods.insert(
-            key.clone(),
-            func_info_from_behavior_method(key, &default.params, &default.return_type),
-        );
-    }
-
-    fn impl_methods_include_behavior_method(
-        &self,
-        type_name: &str,
-        methods: &[Declaration],
-        required_name: &str,
-    ) -> bool {
-        methods
-            .iter()
-            .any(|decl| matches!(decl, Declaration::Function { name, .. } if name == required_name))
-            || (self.resolver_backed_collection
-                && self
-                    .resolver_backed_method_signature(type_name, required_name)
-                    .is_some())
-    }
-
-    fn impl_effective_method_name(
-        &self,
-        unmatched_required: &mut VecDeque<String>,
-        ast_name: &str,
-        resolver_owned_key: Option<String>,
-        type_name: &str,
-    ) -> String {
-        if let Some(resolver_owned_key) = resolver_owned_key {
-            let resolver_owned_name =
-                method_signature_method_name_for_receiver(&resolver_owned_key, type_name)
-                    .unwrap_or(&resolver_owned_key)
-                    .to_string();
-            return Self::remove_named_queue_entry(unmatched_required, &resolver_owned_name)
-                .unwrap_or(resolver_owned_name);
-        }
-
-        if let Some(name) = Self::remove_named_queue_entry(unmatched_required, ast_name) {
-            return name;
-        }
-
-        if self.resolver_backed_collection {
-            if let Some(index) = unmatched_required.iter().position(|required| {
-                self.resolver_backed_method_signature(type_name, required)
-                    .is_some()
-            }) {
-                return unmatched_required
-                    .remove(index)
-                    .unwrap_or_else(|| ast_name.to_string());
-            }
-        }
-
-        ast_name.to_string()
-    }
-
-    fn effective_behavior_impl_methods<'a>(
-        &self,
-        symbols: Option<&SymbolTable>,
-        type_name: &str,
-        methods: &'a [Declaration],
-        unmatched_required: &mut VecDeque<String>,
-    ) -> Vec<EffectiveBehaviorImplMethod<'a>> {
-        methods
-            .iter()
-            .map(|method| {
-                let ast_name = match method {
-                    Declaration::Function { name, .. } => name.as_str(),
-                    _ => "",
-                };
-                let ast_key = Self::method_key(type_name, ast_name);
-                let resolver_owned_name = self.resolver_backed_impl_method_key(
-                    symbols,
-                    &ast_key,
-                    type_name,
-                    method.span(),
-                );
-                let method_name = self.impl_effective_method_name(
-                    unmatched_required,
-                    ast_name,
-                    resolver_owned_name,
-                    type_name,
-                );
-                EffectiveBehaviorImplMethod {
-                    declaration: method,
-                    method_name,
-                }
-            })
-            .collect()
-    }
-
-    fn resolver_backed_behavior_impl_method_signature_name(
-        &self,
-        required_methods: &mut VecDeque<ast::BehaviorMethod>,
-        ast_name: &str,
-        resolver_owned_key: Option<&str>,
-        type_name: &str,
-    ) -> Option<String> {
-        if let Some(resolver_owned_key) = resolver_owned_key {
-            let resolver_owned_name =
-                method_signature_method_name_for_receiver(resolver_owned_key, type_name)
-                    .unwrap_or(resolver_owned_key);
-            if let Some(index) =
-                Self::named_queue_index(required_methods, resolver_owned_name, |required| {
-                    required.name.as_str()
-                })
-            {
-                return required_methods.remove(index).map(|required| required.name);
-            }
-        }
-
-        Self::named_queue_index(required_methods, ast_name, |required| {
-            required.name.as_str()
-        })
-        .and_then(|index| required_methods.remove(index).map(|required| required.name))
-    }
-
-    fn resolver_backed_impl_method_key(
-        &self,
-        symbols: Option<&SymbolTable>,
-        ast_key: &str,
-        type_name: &str,
-        span: Span,
-    ) -> Option<String> {
-        self.resolver_backed_collection
-            .then(|| Self::validation_method_key(symbols, ast_key, type_name, span))
-    }
-
-    fn resolver_backed_method_signature(
-        &self,
-        type_name: &str,
-        method_name: &str,
-    ) -> Option<&FuncInfo> {
-        self.resolver_backed_collection
-            .then(|| self.methods.get(&Self::method_key(type_name, method_name)))
-            .flatten()
-    }
-
-    fn method_key(type_name: &str, method_name: &str) -> String {
-        method_signature_key(type_name, method_name)
-    }
-
-    fn remove_named_queue_entry(items: &mut VecDeque<String>, name: &str) -> Option<String> {
-        items
-            .iter()
-            .position(|item| item == name)
-            .and_then(|index| items.remove(index))
-    }
-
-    fn behavior_methods_with_inherited_substituted(
-        &self,
-        behavior: &str,
-        substitutions: &HashMap<String, AstType>,
-        seen: &mut HashSet<String>,
-    ) -> Vec<ast::BehaviorMethod> {
-        if !self.mark_behavior_seen(behavior, substitutions, seen) {
-            return Vec::new();
-        }
-
-        let mut methods = Vec::new();
-        if let Some(parents) = self.behavior_extends.get(behavior) {
-            for parent in parents {
-                let parent_substitutions =
-                    self.behavior_parent_type_param_substitutions(parent, substitutions);
-                methods.extend(self.behavior_methods_with_inherited_substituted(
-                    &parent.behavior,
-                    &parent_substitutions,
-                    seen,
-                ));
-            }
-        }
-        if let Some(info) = self.behaviors.get(behavior) {
-            methods.extend(
-                info.methods
-                    .iter()
-                    .map(|method| substituted_behavior_method_signature(method, substitutions)),
-            );
-        }
-        methods
-    }
-
-    fn behavior_seen_key(
-        &self,
-        behavior: &str,
-        substitutions: &HashMap<String, AstType>,
-    ) -> String {
-        let type_args = self
-            .behaviors
-            .get(behavior)
-            .map(|info| {
-                info.type_params
-                    .iter()
-                    .filter_map(|param| substitutions.get(param).cloned())
-                    .collect::<Vec<_>>()
-            })
-            .unwrap_or_default();
-        behavior_ref_display(behavior, &type_args)
-    }
-
-    fn mark_behavior_seen(
-        &self,
-        behavior: &str,
-        substitutions: &HashMap<String, AstType>,
-        seen: &mut HashSet<String>,
-    ) -> bool {
-        let behavior_seen_key = self.behavior_seen_key(behavior, substitutions);
-        seen.insert(behavior_seen_key)
-    }
-
-    fn behavior_methods_for_impl(
-        &self,
-        behavior: &str,
-        substitutions: &HashMap<String, AstType>,
-        seen: &mut HashSet<String>,
-    ) -> Vec<ast::BehaviorMethod> {
-        self.behavior_methods_with_inherited_substituted(behavior, substitutions, seen)
-    }
-
-    fn behavior_type_param_substitutions(
-        &self,
-        behavior: &str,
-        type_args: &[AstType],
-    ) -> HashMap<String, AstType> {
-        self.behaviors
-            .get(behavior)
-            .map(|info| {
-                info.type_params
-                    .iter()
-                    .cloned()
-                    .zip(type_args.iter().cloned())
-                    .collect()
-            })
-            .unwrap_or_default()
-    }
-
-    fn behavior_parent_type_param_substitutions(
-        &self,
-        parent: &BehaviorParentRef,
-        substitutions: &HashMap<String, AstType>,
-    ) -> HashMap<String, AstType> {
-        let parent_type_args: Vec<AstType> = parent
-            .type_args
-            .iter()
-            .map(|type_arg| substitute_behavior_ast_type(type_arg, substitutions))
-            .collect();
-        self.behavior_type_param_substitutions(&parent.behavior, &parent_type_args)
-    }
-
-    fn impl_ast_types_compatible(
-        &self,
-        expected: &AstType,
-        actual: &AstType,
-        self_type_name: &str,
-    ) -> bool {
-        match expected {
-            AstType::SelfType => matches!(actual, AstType::Named(name) if name == self_type_name),
-            _ => expected == actual,
-        }
-    }
-
-    fn impl_type_display(&self, ty: &AstType, self_type_name: &str) -> String {
-        match ty {
-            AstType::SelfType => self_type_name.to_string(),
-            _ => ty.display_name(),
-        }
     }
 
     fn validate_generic_bounds(&mut self, type_params: &[ast::TypeParam]) {
